backend/requirements.txt
fastapi
uvicorn
websockets
crewai
langchain
langchain-openai
python-multipart
python-docx
PyPDF2
pdfplumber
pymupdf<1.23  # <-- Limit version to avoid Python 3.13 conflicts
python-magic
pandas
python-dotenv
pydantic[email]
aiofiles
supabase
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
reportlab==4.0.6
